requirements.txt
django
django_debug_toolbar
django_easy_select2
mysqlclient
